+++ src/main/java/com/arangodb/internal/velocypack/VPackSerializers.java
@@ -65,7 +65,7 @@ public class VPackSerializers {
         builder.add(attribute, ValueType.ARRAY);
         builder.add(value.getVersion());
         builder.add(value.getType());
-        builder.add(value.getDatabase());
+        builder.add(value.getDbName().get());
         builder.add(value.getRequestType().getType());
         builder.add(value.getRequest());
         builder.add(ValueType.OBJECT);